requirements.txt

//          Copyright Amber Whitlock aka enigmalea 2021
// Distributed under the Boost Software License, Version 1.0.
//    (See accompanying file LICENSE_1_0.txt or copy at
//          https://www.boost.org/LICENSE_1_0.txt)


discord.py>=1.7.0
git+https://github.com/Rapptz/discord-ext-menus
pytz
python-dotenv>=0.17.1
apscheduler>=3.7.0
ao3_api==2.0.8
disputils
dbots
